testing server api with front end api calls and testing context and server responses.

+++ client/src/apiRoutes/CalendarRoutes.js
@@ -1,15 +1,15 @@
-export const fetchCalendar = async (user, calendar_id) => {
-  const response = await fetch(`/api/calendars/${calendar_id}`, {
-    headers: {
-      Authorization: `bearer ${user.token}`,
-    },
-  });
-
-  return response;
-};
-
 export const fetchCalendars = async (user) => {
-  const response = await fetch("/api/calendars", {
+  if (!user || !user.token) {
+    return new Response(
+      JSON.stringify({ error: "User is not authorized to use this route" }),
+      {
+        status: 401,
+        statusText: "Not Authorized",
+        ok: false,
+      }
+    );
+  }
+  const response = await fetch(`/api/calendars`, {
     headers: {
       Authorization: `bearer ${user.token}`,
     },
@@ -19,7 +19,27 @@ export const fetchCalendars = async (user) => {
 };
 
 export const patchCalendar = async (user, calendar) => {
-  const response = await fetch(`/api/calendars/${calendar.id}`, {
+  if (!user || !user.token) {
+    return new Response(
+      JSON.stringify({ error: "User is not authorized to use this route" }),
+      {
+        status: 401,
+        statusText: "Not Authorized",
+        ok: false,
+      }
+    );
+  }
+  if (!calendar || !calendar.id) {
+    return new Response(
+      JSON.stringify({ error: "Must provide a calendar object to patch" }),
+      {
+        status: 400,
+        statusText: "Bad request",
+        ok: false,
+      }
+    );
+  }
+  const response = await fetch(`/api/calendars/patch`, {
     method: "PATCH",
     headers: {
       Authorization: `bearer ${user.token}`,
@@ -32,7 +52,27 @@ export const patchCalendar = async (user, calendar) => {
 };
 
 export const deleteCalendar = async (user, calendar) => {
-  const response = await fetch(`/api/calendars/${calendar.id}`, {
+  if (!user || !user.token) {
+    return new Response(
+      JSON.stringify({ error: "User is not authorized to use this route" }),
+      {
+        status: 401,
+        statusText: "Not Authorized",
+        ok: false,
+      }
+    );
+  }
+  if (!calendar) {
+    return new Response(
+      JSON.stringify({ error: "Must provide a calendar object to delete" }),
+      {
+        status: 400,
+        statusText: "Bad request",
+        ok: false,
+      }
+    );
+  }
+  const response = await fetch(`/api/calendars/delete/${calendar.id}`, {
     method: "DELETE",
     headers: {
       Authorization: `bearer ${user.token}`,
@@ -43,6 +83,16 @@ export const deleteCalendar = async (user, calendar) => {
 };
 
 export const createCalendar = async (user, calendar) => {
+  if (!user || !user.token) {
+    return new Response(
+      JSON.stringify({ error: "User is not authorized to use this route" }),
+      {
+        status: 401,
+        statusText: "Not Authorized",
+        ok: false,
+      }
+    );
+  }
   const response = await fetch(`/api/calendars`, {
     method: "POST",
     headers: {
